#65 Merge max_count, min_count, and count into a single type
Fix #65
