+++ cmake/build_types/unit_test.cmake
@@ -20,14 +20,12 @@ set(TEST_SRCS
     ${CMAKE_CURRENT_SOURCE_DIR}/parsing_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/custom_parser_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/alias_test.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/policy/count_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/default_value_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/dependent_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/description_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/display_name_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/long_name_test.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/policy/max_count_test.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/policy/min_count_test.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/policy/min_max_count_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/required_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/router_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/policy/short_name_test.cpp
