+++ include/arg_router/policy/validator.hpp
@@ -634,11 +634,9 @@ inline constexpr auto default_validator = validator<
     // one_of
     rule_q<common_rules::despecialised_any_of_rule<dependency::one_of_t>,
            must_not_have_policies<policy::alias_t,
-                                  policy::count_t,
                                   policy::custom_parser,
                                   policy::dependent_t,
-                                  policy::max_count_t,
-                                  policy::min_count_t,
+                                  policy::min_max_count_t,
                                   policy::no_result_value,
                                   policy::router,
                                   policy::validation::validator>,
@@ -649,12 +647,10 @@ inline constexpr auto default_validator = validator<
     rule_q<
         common_rules::despecialised_any_of_rule<mode_t>,
         must_not_have_policies<policy::alias_t,
-                               policy::count_t,
                                policy::custom_parser,
                                policy::default_value,
                                policy::dependent_t,
-                               policy::max_count_t,
-                               policy::min_count_t,
+                               policy::min_max_count_t,
                                policy::multi_stage_value,
                                policy::required_t>,
         positional_args_must_be_at_end<positional_arg_t>,
@@ -665,15 +661,13 @@ inline constexpr auto default_validator = validator<
     rule_q<common_rules::despecialised_any_of_rule<root_t>,
            must_have_policies<policy::validation::validator>,
            must_not_have_policies<policy::alias_t,
-                                  policy::count_t,
                                   policy::custom_parser,
                                   policy::default_value,
                                   policy::dependent_t,
                                   policy::description_t,
                                   policy::display_name_t,
                                   policy::long_name_t,
-                                  policy::max_count_t,
-                                  policy::min_count_t,
+                                  policy::min_max_count_t,
                                   policy::multi_stage_value,
                                   policy::no_result_value,
                                   policy::none_name_t,
